## problem description

`api.get_dataframe` (and therefore `DataFrame.from_deeporigin`) can take extremely long because:

1. 2 calls to `list_database_rows` and `describe_row` take place in series when they could take place in parallel (this is a problem of 700ms vs 300ms)
2. (more serious problem). there is a place where N `DescriebFile` calls are made in series, where N is the number of files in the database. this means that the time it takes to fetch data grows with the number of files linearly, and the time for each file resolution is ~350ms. 

## proposed solution in this PR

using a new `ListFiles` endpoint that allows us to resolve all file IDs and names in a single API call (which typically takes 300ms no matter how many files)

### screenshot

from cold start, this is how long it takes to display a dataframe. in the old version, this took ~10s

<img width="699" alt="Screenshot 2024-08-30 at 8 52 19 AM" src="https://github.com/user-attachments/assets/71b16e2c-ba1c-425a-a8ec-2fd7dce9bf9f">



## rejected alternative

using async methods. this approach was tried in this PR: https://github.com/deeporiginbio/deeporigin-client/pull/73 

it was rejected because it's much slower, and complicates testing dramatically